+++ normtest/__init__.py
@@ -5,4 +5,4 @@ from .version import __name__
 from .version import __version__
 
 
-# from .ryan_joiner import rj_test
+from .ryan_joiner import rj_test