+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### First Release
 
 #### Module ryan_joiner
+- Added ``rj_test`` function (06/11/2023);
 - Added ``_statistic`` function (06/11/2023);
 - Added ``_p_value`` function (06/11/2023);
 - Added ``_order_statistic`` function (06/11/2023);